Moved test suite into seperate module for compatibility on all OS
