+++ tests/model_build_test.py
@@ -7,9 +7,7 @@ from vtk.util import numpy_support as nps
 from vtk.numpy_interface import dataset_adapter as dsa
 
 # Functionality to test:
-from .earth import *
-from .evenModel import *
-from .oddModel import *
+from PVGeo.model_build import *
 
 ###############################################################################
 
@@ -100,4 +98,14 @@ class TestCreateTensorMesh(unittest.TestCase):
         self.assertEqual(bounds, tb)
 
 
+###############################################################################
+
+
+###############################################################################
+###############################################################################
+###############################################################################
+if __name__ == '__main__':
+    unittest.main()
+###############################################################################
+###############################################################################
 ###############################################################################
