+++ tests/readers_test.py
@@ -9,8 +9,7 @@ from vtk.util import numpy_support as nps
 from vtk.numpy_interface import dataset_adapter as dsa
 
 # Functionality to test:
-from .delimited import *
-from .binaries import *
+from PVGeo.readers import *
 
 RTOL = 0.000001
 
@@ -364,5 +363,11 @@ class TestMadagascarReader(unittest.TestCase):
 ###############################################################################
 
 
+###############################################################################
+###############################################################################
+###############################################################################
 if __name__ == '__main__':
     unittest.main()
+###############################################################################
+###############################################################################
+###############################################################################
